Updating parser and task scripts

+++ task_scripts/chunk_ccs.py
@@ -26,13 +26,13 @@ parser.add_argument("sz_thresh", type=int)
 parser.add_argument("--chunk_begin", nargs=3, type=int, required=True)
 parser.add_argument("--chunk_end", nargs=3, type=int, required=True)
 parser.add_argument("--parallel", type=int, default=1)
-parser.add_argument("--mip", nargs="+", type=int, default=(0,))
+parser.add_argument("--mip", nargs="+", type=float, default=(0,))
 parser.add_argument("--hashmax", type=int, default=1)
 parser.add_argument("--timing_tag", default=None)
 
 
 # MIP arguments can specify voxel resolutions or mip index
-def mip_or_res(x): return x[0] if (x is not None and len(x) == 1) else x
+def mip_or_res(x): return int(x[0]) if (x is not None and len(x) == 1) else x
 
 
 args = parser.parse_args()
