+++ task_scripts/chunk_edges.py
@@ -24,15 +24,15 @@ parser.add_argument("--root_seg_cvname", default=None)
 parser.add_argument("--storagedir", default=None, required=True)
 
 # Processing Parameters
-parser.add_argument("--resolution", nargs=3, type=int, default=(4, 4, 40))
+parser.add_argument("--resolution", nargs=3, type=float, default=(4, 4, 40))
 parser.add_argument("--patchsz", nargs=3, type=int, required=True)
 parser.add_argument("--chunk_begin", nargs=3, type=int, required=True)
 parser.add_argument("--chunk_end", nargs=3, type=int, required=True)
 parser.add_argument("--samples_per_cleft", type=int, default=1)
 parser.add_argument("--dil_param", type=int, default=5)
 parser.add_argument("--num_downsamples", type=int, default=0)
-parser.add_argument("--base_res_begin", nargs=3, type=int, default=None)
-parser.add_argument("--base_res_end", nargs=3, type=int, default=None)
+parser.add_argument("--base_res_begin", nargs=3, type=float, default=None)
+parser.add_argument("--base_res_end", nargs=3, type=float, default=None)
 parser.add_argument("--parallel", type=int, default=1)
 parser.add_argument("--timing_tag", default=None)
