+++ task_scripts/chunk_overlaps.py
@@ -21,13 +21,13 @@ parser.add_argument("storagedir")
 parser.add_argument("--chunk_begin", nargs="+", type=int, required=True)
 parser.add_argument("--chunk_end", nargs="+", type=int, required=True)
 parser.add_argument("--parallel", type=int, default=1)
-parser.add_argument("--mip", nargs="+", type=int, default=(0,))
-parser.add_argument("--seg_mip", nargs="+", type=int, default=None)
+parser.add_argument("--mip", nargs="+", type=float, default=(0,))
+parser.add_argument("--seg_mip", nargs="+", type=float, default=None)
 parser.add_argument("--timing_tag", default=None)
 
 
 # MIP arguments can specify voxel resolutions or mip index
-def mip_or_res(x): return x[0] if (x is not None and len(x) == 1) else x
+def mip_or_res(x): return int(x[0]) if (x is not None and len(x) == 1) else x
 
 
 args = parser.parse_args()
