+++ task_scripts/merge_dups.py
@@ -18,7 +18,7 @@ parser.add_argument("hash_index", type=int)
 # Processing Parameters
 parser.add_argument("dist_thr", type=int)
 parser.add_argument("size_thr", type=int)
-parser.add_argument("--voxel_res", nargs=3, type=int, required=True)
+parser.add_argument("--voxel_res", nargs=3, type=float, required=True)
 parser.add_argument("--dst_storagestr", default=None)
 parser.add_argument("--timing_tag", default=None)
